+++ resources/js/Layouts/AdminLayout.jsx
@@ -1,265 +1,21 @@
 import React from 'react';
-import Drawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
-import { ThemeProvider, createTheme } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
+
+
+import AsideDrawer from '@/_Partials/AsideDrawer';
 import Toolbar from '@mui/material/Toolbar';
-import AppBar from '@mui/material/AppBar';
-import List from '@mui/material/List';
-import ListSubheader from '@mui/material/ListSubheader';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-import MenuIcon from '@mui/icons-material/Menu';
 
 
-import Dropdown from '@/Components/Dropdown';
-import ResponsiveNavLink from '@/Components/ResponsiveNavLink';
-
-import translation from '@/../lang/es.json'
-
-const darkTheme = createTheme({
-  palette: {
-    mode: 'dark',
-    background: {
-        default: '#1c232f',
-        paper: '#1c232f',
-    },
-    text: {
-        primary: "#ced4dc",
-        secondary: "#ced4dc",
-    },
-  },
-});
-
-
-const drawerWidth = 280;
-
-export default function AdminLayout({user, children, props})
+export default function AdminLayout({user, children})
 {
-    // const { window } = props;
-    const [mobileOpen, setMobileOpen] = React.useState(false);
-    const [isClosing, setIsClosing] = React.useState(false);
-    const [showingNavigationDropdown, setShowingNavigationDropdown] = React.useState(false);
-
-    const handleDrawerClose = () => {
-        setIsClosing(true);
-        setMobileOpen(false);
-    };
-
-    const handleDrawerTransitionEnd = () => {
-        setIsClosing(false);
-    };
-
-    const handleDrawerToggle = () => {
-        if (!isClosing) {
-            setMobileOpen(!mobileOpen);
-        }
-    };
-
-
-
-
-
-    
-
-    const drawer = (
-        <div>
-            <Box className="h-20 bg-[#161c25] grid place-content-center">
-                <h1 className='text-bold text-3xl uppercase '>Intranet UC</h1>
-            </Box>
-            <Divider />
-            <nav>                            
-                <List 
-                    subheader={
-                        <ListSubheader component="div" id="role-and-permission">
-                            <span className='ml-2 text-[#7267ef] capitalize'>Navegacion</span>
-                        </ListSubheader>
-                    }
-                >
-                    <ListItem disablePadding>
-                        <ListItemButton component="a" href={route('dashboard')} selected={route().current('dashboard')}>
-                            <ListItemText primary="Dashboard"/>
-                        </ListItemButton>
-                    </ListItem>
-                    <ListItem disablePadding>
-                        <ListItemButton component="a" href={route('hola')} selected={route().current('hola')}>
-                            <ListItemText primary="Hola"/>
-                        </ListItemButton>
-                    </ListItem>
-                </List>
-
-
-
-
-                <List
-                    subheader={
-                        <ListSubheader component="div" id="role-and-permission">
-                            <span className='ml-2 text-[#7267ef] capitalize'>roles y permisos</span>
-                        </ListSubheader>
-                    }
-                >
-                    <ListItem disablePadding>
-                        <ListItemButton component="a" href={route('dashboard')} selected={route().current('dashboard')}>
-                            <ListItemText primary="Dashboard"/>
-                        </ListItemButton>
-                    </ListItem>
-                    <ListItem disablePadding>
-                        <ListItemButton component="a" href={route('hola')} selected={route().current('hola')}>
-                            <ListItemText primary="Hola"/>
-                        </ListItemButton>
-                    </ListItem>
-                </List>
-            </nav>
-        </div>
-    )
-
-
-
-
-
+    const drawerWidth = 280;
 
     return(
-        
         <Box sx={{ display: 'flex' }}>
             <CssBaseline />
 
-            <AppBar
-                position="fixed"
-                sx={{
-                    width: { sm: `calc(100% - ${drawerWidth}px)` },
-                    ml: { sm: `${drawerWidth}px` },
-                    background: 'white',
-                }}
-            >
-                <Toolbar className="bg-white h-20 flex justify-between">
-                    <div className='flex items-center'>
-                        <IconButton
-                            color="black"
-                            aria-label="open drawer"
-                            edge="start"
-                            onClick={handleDrawerToggle}
-                            sx={{ mr: 2, display: { sm: 'none' } }}
-                        >
-                            <MenuIcon />
-                        </IconButton>
-                        <p className='text-black text-lg font-bold'>Tool Bar</p>
-                    </div>
-
-                    <nav>
-                        <div className="hidden sm:flex sm:items-center sm:ms-6">
-                            <div className="ms-3 relative">
-                                <Dropdown>
-                                    <Dropdown.Trigger>
-                                        <span className="inline-flex rounded-md">
-                                            <button
-                                                type="button"
-                                                className="inline-flex items-center px-3 py-2 border border-transparent text-sm leading-4 font-medium rounded-md text-gray-500 bg-white hover:text-gray-700 focus:outline-none transition ease-in-out duration-150"
-                                            >
-                                                {user.name}
-
-                                                <svg
-                                                    className="ms-2 -me-0.5 h-4 w-4"
-                                                    xmlns="http://www.w3.org/2000/svg"
-                                                    viewBox="0 0 20 20"
-                                                    fill="currentColor"
-                                                >
-                                                    <path
-                                                        fillRule="evenodd"
-                                                        d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
-                                                        clipRule="evenodd"
-                                                    />
-                                                </svg>
-                                            </button>
-                                        </span>
-                                    </Dropdown.Trigger>
-
-                                    <Dropdown.Content>
-                                        <Dropdown.Link href={route('profile.edit')}>{translation['Profile']}</Dropdown.Link>
-                                        <Dropdown.Link href={route('logout')} method="post" as="button">
-                                            {translation['Log Out']}
-                                        </Dropdown.Link>
-                                    </Dropdown.Content>
-                                </Dropdown>
-                            </div>
-                        </div>
-
-                        <div className="-me-2 flex items-center sm:hidden">
-                            <button
-                                onClick={() => setShowingNavigationDropdown((previousState) => !previousState)}
-                                className="inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:bg-gray-100 focus:text-gray-500 transition duration-150 ease-in-out"
-                            >
-                                <svg className="h-6 w-6" stroke="currentColor" fill="none" viewBox="0 0 24 24">
-                                    <path
-                                        className={!showingNavigationDropdown ? 'inline-flex' : 'hidden'}
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        strokeWidth="2"
-                                        d="M4 6h16M4 12h16M4 18h16"
-                                    />
-                                    <path
-                                        className={showingNavigationDropdown ? 'inline-flex' : 'hidden'}
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        strokeWidth="2"
-                                        d="M6 18L18 6M6 6l12 12"
-                                    />
-                                </svg>
-                            </button>
-                        </div>
-                    </nav>                   
-                </Toolbar>
-                <div className={(showingNavigationDropdown ? 'block' : 'hidden') + ' sm:hidden '}>
-                        <div className="pt-4 pb-1 border-t border-gray-200">
-                            <div className="px-4">
-                                <div className="font-medium text-base text-gray-800">{user.name}</div>
-                                <div className="font-medium text-sm text-gray-500">{user.email}</div>
-                            </div>
-
-                            <div className="mt-3 space-y-1">
-                                <ResponsiveNavLink href={route('profile.edit')}>{translation['Profile']}</ResponsiveNavLink>
-                                <ResponsiveNavLink method="post" href={route('logout')} as="button">
-                                    {translation['Log Out']}
-                                </ResponsiveNavLink>
-                            </div>
-                        </div>
-                    </div>
-            </AppBar>
-
-            <ThemeProvider theme={darkTheme}>
-                <Drawer
-                    sx={{
-                        display: { xs: 'block', sm: 'none' },
-                        
-                        '& .MuiDrawer-paper': { boxSizing: 'border-box', width: drawerWidth },
-                    }}
-                    variant="temporary"
-                    anchor="left"
-                    open={mobileOpen}
-                    onTransitionEnd={handleDrawerTransitionEnd}
-                    onClose={handleDrawerClose}
-                    ModalProps={{
-                        keepMounted: true, // Better open performance on mobile.
-                    }}
-                >
-                    {drawer}
-                </Drawer>
-                <Drawer
-                    variant="permanent"
-                    anchor="left"
-                    sx={{
-                        display: { xs: 'none', sm: 'block' },
-                        width: drawerWidth,
-                        flexShrink: 0,
-                        '& .MuiDrawer-paper': { boxSizing: 'border-box', width: drawerWidth },
-                    }}
-                    open
-                    >
-                        {drawer}
-                </Drawer>
-            </ThemeProvider>
+            <AsideDrawer user={user} drawerWidth={drawerWidth}/>
             
             <Box
                 component="main"
